feat(add): Add new SVG images and update backend routes

+++ frontend/src/components/Shop/SingleShop.tsx
@@ -1,14 +1,28 @@
 import React, {CSSProperties, useEffect, useState} from 'react';
 import {useNavigate, useParams} from 'react-router-dom';
 import "./form.scss"
-import {Autocomplete, Button, createFilterOptions, TextField, Tooltip} from "@mui/material";
+import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
+import {
+    Accordion,
+    AccordionDetails,
+    AccordionSummary,
+    Autocomplete,
+    Chip,
+    createFilterOptions,
+    TextField,
+    Tooltip
+} from "@mui/material";
 import {Product} from "../Product/Product";
 import {ProductModel} from "../../models/Product.model";
 import randomColor from "randomcolor";
 import {ShopModel} from "../../models/Shop.model";
 import {ShopService} from "../../services/shop.service";
-import DeleteIcon from '@mui/icons-material/Delete';
 import {ProductService} from "../../services/product.service";
+import remove from '../../images/trash-bin-minimalistic-svgrepo-com.svg';
+import plein from '../../images/forbidden-circle-svgrepo-com.svg';
+import vide from '../../images/record-svgrepo-com.svg';
+import rayon from '../../images/bedside-table-2-svgrepo-com.svg';
+import beacon from '../../images/vinyl-record-svgrepo-com.svg';
 
 export enum PlanType {
     VIDE = "VIDE",
@@ -62,8 +76,11 @@ const Plan = (props: CustomProps) => {
 
     const filter = createFilterOptions<string>();
 
-    const handleChangeRadio = (event: any) => {
-        setCurrentDraw(event.target.value as PlanType);
+    const [selectedRayonList, setSelectedRayonList] = React.useState<string | null>(rayons[0]);
+    const [produitRayon, setProduitayon] = useState<Array<ProductModel>>([]);
+
+    const handleChangeRadio = (planType: PlanType) => {
+        setCurrentDraw(planType)
     };
 
     const [info, setInfo] = useState<{ open: boolean, message: string, type: "success" | "error" }>({
@@ -230,6 +247,7 @@ const Plan = (props: CustomProps) => {
 
         ProductService.addNewProductToRayonP({storeId: props.storeId, productId: product.id, name: value}).then(response => {
             console.log(response);
+            setInfo({open: true, message: "L'affection a été enregistrée !", type: "success"});
         })
     }
 
@@ -309,28 +327,50 @@ const Plan = (props: CustomProps) => {
         setColorRayons(rayons.length)
     }
 
+    const renderTooltipWithImage = (typeDePlan: PlanType, index: number) => {
+        const images: any = {
+            [PlanType.BEACON]: beacon,
+            [PlanType.RAYON]: rayon,
+            [PlanType.VIDE]: vide,
+            [PlanType.PLEIN]: plein,
+        };
+
+        return (
+            <Tooltip title={typeDePlan.toLowerCase()} key={index}>
+                <img className={currentDraw === typeDePlan ? "click" : ""} src={images[typeDePlan]} alt={typeDePlan}
+                     onClick={() => handleChangeRadio(typeDePlan)}/>
+            </Tooltip>
+        );
+    };
+
+    const handleClickProduitRayon = (data: any) => {
+        console.log(data)
+    }
+
+    const getProduitRayon = (rayon: string) => {
+        ProductService.getProductsBySectionName(rayon).then(response => {
+            setProduitayon(response.map((tmp: any) => tmp.product))
+        })
+    }
+
     return (
         <div className={"single_shop"}>
+
             {info.open ? <div className={"info " + info.type}>
                 <p>{info.message}</p>
                 <div onClick={() => updateOpen(false)}>X</div>
             </div> : ""}
+
             <div className={"configPlan"}>
                 <div>
                     <h3>Choix du stylo</h3>
-                    <div>
-                        {getListDrawType().map((planType, index) => (
-                            <label key={index} className={"type"}>
-                                <input
-                                    type="radio"
-                                    name={"checkbox"}
-                                    value={planType}
-                                    checked={currentDraw === planType}
-                                    onChange={handleChangeRadio}
-                                />
-                                {planType}
-                            </label>
-                        ))}
+                    <div className={"radio"}>
+                        {getListDrawType().map((planType, index) => {
+                            if ([PlanType.BEACON, PlanType.RAYON, PlanType.VIDE, PlanType.PLEIN].includes(planType)) {
+                                return renderTooltipWithImage(planType, index);
+                            }
+                            return null;
+                        })}
                     </div>
                 </div>
                 <div>
@@ -439,6 +479,39 @@ const Plan = (props: CustomProps) => {
                 })}
                 </tbody>
             </table>
+
+            <Accordion>
+                <AccordionSummary
+                    expandIcon={<ExpandMoreIcon/>}
+                    aria-controls="panel1-content"
+                    id="panel1-header"
+                >
+                    <h2>Liste les produits par rayon</h2>
+                </AccordionSummary>
+                <AccordionDetails>
+                    <div>
+                        <div>
+                            <Autocomplete
+                                value={selectedRayonList}
+                                onChange={(event: any, newValue: string | null) => {
+                                    if (newValue) {
+                                        setSelectedRayonList(newValue);
+                                        getProduitRayon(newValue);
+                                    }
+                                }}
+                                id="controllable-states-demo"
+                                options={rayons}
+                                sx={{width: 300}}
+                                renderInput={(params) => <TextField {...params} label="Liste des rayons"/>}
+                            />
+                        </div>
+                        <div>
+                            {produitRayon.map((product: ProductModel) => <Chip label={product.name}
+                                                                               onClick={handleClickProduitRayon}/>)}
+                        </div>
+                    </div>
+                </AccordionDetails>
+            </Accordion>
         </div>
     )
 }
@@ -473,58 +546,16 @@ const SingleShop: React.FC = () => {
     const navigate = useNavigate();
 
     const removeShop = (id: number) => {
-        ShopService.removePlan(id).then(() => navigate(`/shop`));
+        if (window.confirm(`Souhaitez-vous vraiment supprimer le ${shop?.name} ?`)) {
+            ShopService.removePlan(id).then(() => navigate(`/shop`));
+        }
     }
 
     return (!shop ? <div>Loading...</div> :
-        <div>
-            <div className={"topContent"}>
-                <div>
-                    <h2>Id du Shop</h2>
-                    <h3>{shop.id}</h3>
-                </div>
-                <div>
-                    <h2>Nom</h2>
-                    <TextField
-                        id="outlined-controlled"
-                        label="Nombres de rows"
-                        value={shop.name}
-                        type={"text"}
-                        onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-                            setShop({...shop, name: event.target.value});
-                        }}
-                    />
-                </div>
-                <div>
-                    <h2>Adresse</h2>
-                    <TextField
-                        id="outlined-controlled"
-                        label="Nombres de rows"
-                        value={shop.adresse}
-                        type={"text"}
-                        onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-                            setShop({...shop, adresse: event.target.value});
-                        }}
-                    />
-                </div>
-                <div>
-                    <h2>Ville</h2>
-                    <TextField
-                        id="outlined-controlled"
-                        label="Nombres de rows"
-                        value={shop.ville}
-                        type={"text"}
-                        onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-                            setShop({...shop, ville: event.target.value});
-                        }}
-                    />
-                </div>
-                <div>
-                    <h2>Suppression</h2>
-                    <Button color={"error"} onClick={() => removeShop(shop?.id)} variant="outlined" startIcon={<DeleteIcon />}>
-                        Delete
-                    </Button>
-                </div>
+            <div className={"singleShop"}>
+                <div className={"top"}>
+                    <h1>Hi, bienvenue sur {shop.name} 👋</h1>
+                    <img src={remove} alt="remove icon trash" onClick={() => removeShop(shop?.id)}/>
             </div>
             <Plan name={shop.name} storeId={shop.id} onFinish={updatePlan} plan={shop.layout} isEdit={isEdit}/>
         </div>
